+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Fixed
+* `flexberry-simpleolv`, `flexberry-objectlistview`, `object-list-view` components:
+    * Fix `showLoadingTbodyClass` event.
 * Component `flexberry-lookup` override private property `context` from `Ember.Component` that could cause errors.
 
 ## [0.7.0-beta.15] - 2016-10-10